+++ lib/ReactViews/Credits/Leaflet/LeafletCredits.props.ts
@@ -1,0 +1,8 @@
+import { ICredit } from "../Credit.type";
+
+export interface ILeafletCreditsProps {
+  hideTerriaLogo: boolean;
+  prefix?: string;
+  credits?: ICredit[];
+  dataAttributions: string[];
+}