+++ lib/ReactViews/Credits/index.ts
@@ -1,0 +1,2 @@
+export { CesiumCredits } from "./Cesium/CesiumCredits";
+export { LeafletCredits } from "./Leaflet/LeafletCredits";